+++ EntidadesH/ObraSocial.h
@@ -6,20 +6,20 @@ class ObraSocial{
     private:
         int _id;
         char _nombre [30];
-        int _descuento;
+        float _descuento;
         bool _estado;
 
     public:
         ObraSocial();
-        ObraSocial(int id, std::string nombre, int descuento);
+        ObraSocial(int id, std::string nombre, float descuento);
 
         int getId() const;
         std::string getNombre() const;
-        int getDescuento() const;
+        float getDescuento() const;
         bool getEstado() const;
 
         void setId(int idObraSocial);
         void setNombre(std::string& nombre);
-        void setDescuento(int descuento);
+        void setDescuento(float descuento);
         void setEstado(bool estado);
 };
